LOOK-244 added DAGs test versions and remove junk
