+++ update_everything_new_dev_test.py
@@ -27,7 +27,7 @@ redshift_conn = redshift_hook.get_conn()
 cursor = redshift_conn.cursor()
 
 # set up environment
-schema = 'staging'
+schema = 'mock'
 
 
 def start_slack_alert(context):
@@ -215,29 +215,29 @@ def upsert_warehouse_order_items(schema, table, date_column, **kwargs):
 
 def get_tasks():
     return [
-        {'task_id': 'upsert_brands', 'op_args': ['brands']},
-        {'task_id': 'upsert_company_config', 'op_args': ['company_config']},
-        {'task_id': 'upsert_discounts', 'op_args': ['discounts']},
-        {'task_id': 'upsert_patient_group_ref', 'op_args': ['patient_group_ref']},
-        {'task_id': 'upsert_patient_group', 'op_args': ['patient_group']},
+        # {'task_id': 'upsert_brands', 'op_args': ['brands']},
+        # {'task_id': 'upsert_company_config', 'op_args': ['company_config']},
+        # {'task_id': 'upsert_discounts', 'op_args': ['discounts']},
+        # {'task_id': 'upsert_patient_group_ref', 'op_args': ['patient_group_ref']},
+        # {'task_id': 'upsert_patient_group', 'op_args': ['patient_group']},
         {'task_id': 'upsert_patients', 'op_args': ['patients']},
-        {'task_id': 'upsert_product_categories', 'op_args': ['product_categories']},
+        # {'task_id': 'upsert_product_categories', 'op_args': ['product_categories']},
         {'task_id': 'upsert_product_checkins', 'op_args': ['product_checkins']},
-        {'task_id': 'upsert_product_filter_index', 'op_args': ['product_filter_index']},
-        {'task_id': 'upsert_product_office_qty', 'op_args': ['product_office_qty']},
-        {'task_id': 'upsert_product_transactions', 'op_args': ['product_transactions']},
-        {'task_id': 'upsert_product_vendors', 'op_args': ['product_vendors']},
+        # {'task_id': 'upsert_product_filter_index', 'op_args': ['product_filter_index']},
+        # {'task_id': 'upsert_product_office_qty', 'op_args': ['product_office_qty']},
+        # {'task_id': 'upsert_product_transactions', 'op_args': ['product_transactions']},
+        # {'task_id': 'upsert_product_vendors', 'op_args': ['product_vendors']},
         {'task_id': 'upsert_products', 'op_args': ['products']},
-        {'task_id': 'upsert_register_log', 'op_args': ['register_log']},
+        # {'task_id': 'upsert_register_log', 'op_args': ['register_log']},
         {'task_id': 'upsert_register', 'op_args': ['register']},
         {'task_id': 'upsert_service_history', 'op_args': ['service_history']},
-        {'task_id': 'upsert_sf_guard_group', 'op_args': ['sf_guard_group']},
-        {'task_id': 'upsert_sf_guard_user_group', 'op_args': ['sf_guard_user_group']},
+        # {'task_id': 'upsert_sf_guard_group', 'op_args': ['sf_guard_group']},
+        # {'task_id': 'upsert_sf_guard_user_group', 'op_args': ['sf_guard_user_group']},
         {'task_id': 'upsert_sf_guard_user', 'op_args': ['sf_guard_user']},
-        {'task_id': 'upsert_sf_guard_user_permission', 'op_args': ['sf_guard_user_permission']},
-        {'task_id': 'upsert_tax_payment', 'op_args': ['tax_payment']},
-        {'task_id': 'upsert_user_activity_record', 'op_args': ['user_activity_record']},
-        {'task_id': 'upsert_warehouse_order_logs', 'op_args': ['warehouse_order_logs']},
+        # {'task_id': 'upsert_sf_guard_user_permission', 'op_args': ['sf_guard_user_permission']},
+        # {'task_id': 'upsert_tax_payment', 'op_args': ['tax_payment']},
+        # {'task_id': 'upsert_user_activity_record', 'op_args': ['user_activity_record']},
+        # {'task_id': 'upsert_warehouse_order_logs', 'op_args': ['warehouse_order_logs']},
         {'task_id': 'upsert_warehouse_orders', 'op_args': ['warehouse_orders']}
     ]
 
@@ -253,10 +253,10 @@ default_args = {
 
 
 with DAG(
-    dag_id='update_everything_new_dev',
+    dag_id='update_everything_new_dev_test',
     max_active_tasks=32,
-    schedule='0 8 * * *', # UTC time
-    start_date=datetime(year=2022, month=12, day=8),
+    schedule=None,  #'0 8 * * *', # UTC time
+    start_date=datetime(year=2023, month=3, day=24),
     default_args=default_args,
     catchup=False,
 ) as dag:
